Extracted Race interfance for all algorithms
Broke up all the algorithms to a Runner and Race implementors and so
they can be accessed throught the common Race interface.

I have also added a Main class that can be run with a parameter
specifying the desired algorithm for this run. This main class will
verify the race result with RaceVerifier and exit with code 1 when
verification is false.